app/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.5
urllib3==2.5.0
flask==3.1.2
Werkzeug==3.1.2
psycopg[binary]
python-dotenv
supabase==2.24.0
flask_dropzone
re